Refactor: Extract VectorIndexBuilder and move vector/temporal tests
- Extracted `VectorIndexBuilder` from `src/db.rs` to `src/api/vector_builder.rs` to reduce coupling and file size.
- Moved Vector Index API tests and Temporal API tests from `src/db.rs` (unit tests) to `tests/vector_api_tests.rs` and `tests/temporal_api_tests.rs` (integration tests).
- Added `__test_current_timestamp` helper to `GallifreyDB` to facilitate integration testing of time-travel queries.
- Cleaned up `src/db.rs` reducing it by ~2000 lines.
- No logic changes, purely refactoring.
